rewrite operators

+++ operators/fbx_transfer.py
@@ -1,10 +1,8 @@
 import bpy
-import time
 
 from ..utils import file_handling
 from ..utils.server_socket import ServerSocket
 from ..utils.csc_handling import CascadeurHandler
-from . import reciever
 
 
 def import_fbx(file_path: str) -> list:
@@ -75,8 +73,9 @@ class CBB_OT_import_cascadeur_fbx(bpy.types.Operator):
                 file_handling.delete_file(data)
                 return {"FINISHED"}
         else:
-            # Report warning!
-            pass
+            self.server_socket.close()
+            self.server_socket = None
+            self.report({"ERROR"}, "No client found.")
 
         return {"PASS_THROUGH"}
 
@@ -93,6 +92,8 @@ class CBB_OT_import_action_to_selected(bpy.types.Operator):
     bl_idname = "cbb.import_cascadeur_action"
     bl_label = "Import Cascadeur Action"
 
+    ao = None
+
     @classmethod
     def poll(cls, context):
         return (
@@ -101,25 +102,39 @@ class CBB_OT_import_action_to_selected(bpy.types.Operator):
             and context.active_object.type == "ARMATURE"
         )
 
-    def execute(self, context):
-        ao = bpy.context.active_object
-        CascadeurHandler().execute_csc_command("commands.externals.temp_exporter.py")
-        bpy.ops.cbb.start_server()
-        data = reciever.recieved_data
-        time.sleep(2)
-        if data:
-            file_handling.wait_for_file(data)
-            imported_objects = import_fbx(data)
-            file_handling.delete_file(data)
-            reciever.recieved_data = None
-        else:
-            print("No data recieved")
+    def modal(self, context, event):
+        if event.type == "ESC":
+            self.server_socket.close()
+            self.server_socket = None
             return {"CANCELLED"}
 
-        actions = get_actions_from_objects(imported_objects)
-        apply_action(ao, actions)
-        delete_objects(imported_objects)
-        return {"FINISHED"}
+        self.server_socket.run()
+
+        if self.server_socket.client_socket:
+            data = self.server_socket.receive_message()
+            if data:
+                print(str(data))
+                file_handling.wait_for_file(data)
+                imported_objects = import_fbx(data)
+                file_handling.delete_file(data)
+                actions = get_actions_from_objects(imported_objects)
+                apply_action(self.ao, actions)
+                delete_objects(imported_objects)
+                return {"FINISHED"}
+        else:
+            self.server_socket.close()
+            self.server_socket = None
+            self.report({"ERROR"}, "No client found.")
+            return {"CANCELLED"}
+
+        return {"PASS_THROUGH"}
+
+    def execute(self, context):
+        self.ao = bpy.context.active_object
+        self.server_socket = ServerSocket()
+        CascadeurHandler().execute_csc_command("commands.externals.temp_exporter.py")
+        context.window_manager.modal_handler_add(self)
+        return {"RUNNING_MODAL"}
 
 
 # Should be moved to a different place
